Added Clear Data and Notification Style button
The Clear Data button, while cosmetically existing
before, had no true functionality. We'll use the
built-in APIs to clear package data for those that
support it, and hackily clean the package using
shell commands for all others.

And for Oreo users, a Notification Style button
for jumping to the app's notification channel
settings in the settings app.