Added current status widget
Added a homescreen/lockscreen widget which shows
the current PRT status. Will likely expand in the
future so that long messages are broken by an
ellipsis and users are able to choose between a
full status message or a simplified "The PRT is X"
message.

Maybe also add a style to display current status
and weather?